refactor: separate out slide properties
