+++ frontend/src/utils/constants.js
@@ -1,0 +1,4 @@
+const sectionClasses = 'flex flex-col gap-4 p-4 border-b'
+const sectionTitleClasses = 'text-2xs font-semibold uppercase text-gray-700'
+
+export { sectionClasses, sectionTitleClasses }
